feat
Write a test that loops through each feed in the allFeeds object and
ensures it has a URL defined and that the URL is not empty.
Write a test that loops through each feed in the allFeeds object and
ensures it has a name defined and that the name is not empty.